### 2025-04-02 #7915 @kthyng
> Hi [@dsheng1026](https://github.com/dsheng1026) — can you link to your previous submission in JOSS?

Ok it is probably not a good sign I haven't heard back about that link, but in rereading the submission cover letter info it looks like I misread a note about a different paper, so I will move on.
