### 2025-04-03 #7915 @dsheng1026
> > Hi [@dsheng1026](https://github.com/dsheng1026) — can you link to your previous submission in JOSS?
>
> Ok it is probably not a good sign I haven't heard back about that link, but in rereading the submission cover letter info it looks like I misread a note about a different paper, so I will move on.

@kthyng Apologies for the late reply! I was out of the office for a while. Thanks for following up, and I’m glad you have sorted it out. I'll make sure to address the comments from editorialbot as soon as possible.
