+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={63},
   pages={102087},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.gloenvcha.2020.102087}
 }
 
 @article{de2021heat,
@@ -16,12 +17,13 @@
   number={4},
   pages={044020},
   year={2021},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/1748-9326/abeb9f}
 }
 
 @misc{shengomitting,
-	title = {Omitting labor responses to heat stress underestimates future climate impact on agriculture},
-	doi = {https://doi.org/10.21203/rs.3.rs-5000229/v1},
+	title = {Omitting labor responses underestimates the effects of future heat stress on agriculture},
+	doi = {10.21203/rs.3.rs-5000229/v1},
 	year={2025},
 	author = {Sheng, Di and Zhao, Xin and Edmonds, James and Waldhoff, Stephanie and Patel, Pralit and O'Neill, Brian and Tebaldi, Claudia and Wise, Marshall},
 }
@@ -33,7 +35,8 @@
   volume={20},
   pages={1--14},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10113-020-01625-6}
 }
 
 @article{spangler2022wet,
@@ -44,13 +47,15 @@
   number={1},
   pages={326},
   year={2022},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41597-022-01405-3}
 }
 
 @article{hijmans2023package,
   title={Package ‘meteor’},
   author={Hijmans, Robert J and Nelson, Gerald},
-  year={2023}
+  year={2023},
+  url={https://cran.r-project.org/web/packages/meteor/index.html}
 }
 
 @article{smallcombe2022quantifying,
@@ -61,7 +66,8 @@
   number={12},
   pages={2463--2476},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00484-022-02370-7}
 }
 
 @article{foster2021advanced,
@@ -71,7 +77,8 @@
   volume={65},
   pages={1215--1229},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00484-021-02105-0}
 }
 
 @article{jagermeyr2021ggcmi,
@@ -79,12 +86,13 @@
   author={J{\"a}germeyr, Jonas and M{\"u}ller, Christoph and Minoli, Sara and Ray, Deepak and Siebert, Stefan},
   journal={(No Title)},
   year={2021},
-  publisher={Zenodo}
+  publisher={Zenodo},
+  doi={10.5281/zenodo.5062513}
 }
 
 @misc{SPAM2024,
   title={Global spatially-disaggregated crop production statistics data for 2020 version 1.0.0},
-  doi = {https://doi.org/10.7910/DVN/SWPENT},
+  doi = {10.7910/DVN/SWPENT},
   author={IFPRI},
   journal={Harvard Dataverse, v1},
   year={2024},
@@ -99,7 +107,8 @@
   number={3},
   pages={e2020EF001885},
   year={2021},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2020EF001885}
 }
 
 @article{kong2022explicit,
@@ -110,7 +119,8 @@
   number={3},
   pages={e2021EF002334},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2021EF002334}
[…]
