### 2025-05-03 1c6aeea
include the harvested area info into G2R results
