### 2025-05-03 a90ece2
include the harvested area info into G2R results
