### 2025-05-26 #7915 @anastassiavybornova
@nmstreethran @jamesdamillington @cjerzak would you be willing to review this submission to JOSS? your input would be greatly appreciated!

if this is your first time around, you can find the reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

-- and for any questions and/or confirming that you would be happy to review the submission, just comment (tagging me) below.
