### 2025-05-28 #7915 @cjerzak
@anastassiavybornova  I'm happy to help review this submission, if this is still needed. This is my first time reviewing for JOSS so bear with me as I get my bearings here!
