### 2025-06-16 #8415 @editorialbot
[REVIEW]: HELPS: an R package to project future Heat Effects on Labor Productivity by Sector
**Submitting author:** @dsheng1026 (<a href="http://orcid.org/0009-0006-7567-380X">Di Sheng</a>)
**Repository:** https://github.com/JGCRI/HELPS
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.0.0
