### 2025-06-16 #8415 @anastassiavybornova
Dear Connor, Nithiya & James @cjerzak @nmstreethran @jamesdamillington ,

huge thanks already for agreeing to review this submission to JOSS!
Apologies for the delay on my side in assigning reviewer roles - since July is coming closer, I'm happy to count on your pair of eyes as well @jamesdamillington :)

instructions on next steps for you as reviewers (starting by generating the checklist) are already posted by the editorialbot in the comments above, and can also be found here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html# . ping me (@anastassiavybornova ) if you need help or clarifications with the reviewing process; and ping the submitting author, Di Sheng @dsheng1026 if you have questions / comments along the way.
