### 2025-07-24 #8415 @anastassiavybornova
@jamesdamillington Thanks for the review so far! to answer your question, yes indeed, the "community guidelines" section here: https://joss.readthedocs.io/en/latest/review_criteria.html outlines that there should be clear guidelines for people wishing to contribute to the software, so @dsheng1026 please do include that (this doesn't have to be very long or elaborate, as long as it is clear!) - regarding the "seek support" part - in this case a contact to one or several of the main developers is enough.
