### 2025-09-28 b75f2a4
add community guideline section
