### 2025-09-28 f524c58
add a brief statement of need in the READEME section
