### 2025-09-28 30a5777
update the reference for package application and GCAM
