+++ paper/paper.md
@@ -44,7 +44,7 @@ Rising evidence of the adverse impact of heat stress on labor [@orlov2020economi
 
 To our knowledge, three existing R packages, HeatStress [@casanueva2020escalating], heatmetrics [@spangler2022wet], and meteor [@hijmans2023package], deliver related outputs. These packages provide functions to calculate heat stress levels using various heat stress metrics, i.e., WBGT, Universal Thermal Climate Index (UTCI), humidex, and heat index. The meteor package further translates heat stress to PWC using the labor-heat response functions from @smallcombe2022quantifying and @foster2021advanced. However, their estimates represent a single set of impacts that do not account for variations in labor heat exposure across crop types and management practices and need further processing to meet the spatial and temporal resolution demands of some research and modeling applications.
 
-HELPS fills this gap by introducing relevant datasets for meaningful temporal and spatial aggregation. Exposure of agricultural labor to heat stress is subject to the time and location of production activities. HELPS integrates crop calendars [@jagermeyr2021ggcmi] and the latest global harvested area data from the Spatial Production Allocation Model (SPAM) data [@SPAM2024] to align the measurements of heat stress and PWC measurements to observed agricultural practices. HELPS allows users to generate output for relevant administrative or environmental boundaries, enhancing its usability across research and modeling applications. HELPS estimates distinguish between irrigated and rainfed systems for 46 SPAM crops and can be expanded to other crops with available calendars and harvested area data. This granularity ensures that HELPS’s outputs capture the most meaningful heat stress impacts on crop labor across regions, management practices, and crop commodities. With uncertainty in heat-induced PWC losses introduced by the choices of heat stress metrics [@buzan2020moist; @schwingshackl2021heat; @kong2022explicit] and labor-heat response functions [@smallcombe2022quantifying; @foster2021advanced], HELPS offers default heat stress function (HS) and LHR functions while also allowing users to incorporate customized functions, ensuring the package can provide output aligning with both established methods and can keep up with future literature advancement. In general, HELPS contributes to the community by providing flexibility of function choices to quantify heat stress and PWC, highlighting the crop-specific measurement that incorporates crop calendar and harvested area data, at various temporal and spatial resolutions to accommodate different research needs.
[…]
