+++ paper/paper.md
@@ -36,7 +36,7 @@ affiliations:
 
 # Summary
 
-Labor productivity, especially in outdoor sectors like primary agriculture, is highly sensitive to weather variations and heat stress. Understanding how labor productivity responds to key heat stress factors, such as temperature and humidity, is essential for evaluating future climate scenarios and their multisectoral dynamic (MSD) impacts. However, global economic and MSD modeling often operate at more aggregate temporal and spatial resolutions than those of climate projections. This resolution disparity highlights the need for tools to connect the fine-scale insights of climate models with the more aggregate scales of economic models. The HELPS R package is designed to generate high-quality data on biophysical labor productivity losses due to heat stress, tailored for modeling applications. HELPS processes high-resolution atmospheric data to calculate heat stress metrics, such as Wet-Bulb Globe Temperature (WBGT) and translates them into physical work capacity (PWC) estimates across various scenarios. The package further provides functions to aggregate data temporally and spatially, incorporating crop calendars and harvested area information to provide relevant measures of heat-induced PWC losses for labor in agricultural sectors. We demonstrate the application of HELPS by generating regional heat-induced agricultural labor productivity losses for major crop sectors, which serve as input data for the Global Change Analysis Model (GCAM), enabling more comprehensive evaluations of climate impacts on agriculture systems. While HELPS is designed to support MSD modeling needs, users can customize output to benefit a broad set of research applications.
+Labor productivity, especially in outdoor sectors like primary agriculture, is highly sensitive to weather variations and heat stress. Understanding how labor productivity responds to key heat stress factors, such as temperature and humidity, is essential for evaluating future climate scenarios and their multisectoral dynamic (MSD) impacts. However, global economic and MSD modeling often operate at more aggregate temporal and spatial resolutions than Earth System Models. This resolution disparity highlights the need for tools to connect the fine-scale insights of Earth System Models with the more aggregate scales of economic models. The HELPS R package is designed to generate high-quality data on biophysical labor productivity losses due to heat stress, tailored for modeling applications. HELPS processes high-resolution atmospheric data to calculate heat stress metrics, such as Wet-Bulb Globe Temperature (WBGT) and translates them into physical work capacity (PWC) estimates across various scenarios. The package further provides functions to aggregate data temporally and spatially, incorporating crop calendars and harvested area information to provide relevant measures of heat-induced PWC losses for labor in agricultural sectors. We demonstrate the application of HELPS by generating regional heat-induced agricultural labor productivity losses for major crop sectors, which serve as input data for the Global Change Analysis Model (GCAM), enabling more comprehensive evaluations of climate impacts on agriculture systems. While HELPS is designed to support MSD modeling needs, users can customize output to benefit a broad set of research applications.
 
 # Statement of need
 
@@ -48,9 +48,9 @@ HELPS fills this gap by introducing relevant datasets for meaningful temporal an
 
 # Features
 
-![HELPS package schematic. HELPS can process daily and monthly 0.5-degree grid-level input data. Stars denote package functions.](Schematic.jpg)
+![HELPS package schematic. HELPS processes daily and monthly input data at 0.5-degree grid resolution. Stars denote package functions.](Schematic.jpg)
 
[…]
