### 2025-09-28 #8415 @dsheng1026
My apologies for the slower turnaround on this revision @anastassiavybornova , and I’ve been out of office for an international relocation. Thank you all for your patience and thoughtful contributions to the review process, and I’m glad to now share the updates. Below is a point-by-point response to the revision comments in addition to the recent commits.

@jamesdamillington
The author contributions are as follows:

Di Sheng: led package development, conceptualized the workflow, and conducted package testing

Xin Zhao: conceptualized the workflow

Stephanie Waldhoff: project management

Abigail Snyder: package testing and development guidance

Chris Vernon: package testing and development guidance

I have also added a brief summary of the statement of need in the Introduction section and included community guidelines in the README.

@nmstreethran
I have incorporated the following updates:

Added community guidelines to the README

Expanded LHR to its full name in paper.md

Aligned installation guidance across the main and paper branches

Corrected the LHR_NIOSH typo and fixed the Step 1.2 rendering issue

Thank you for the careful review and detailed checks!

Regarding your additional points:

In our paper, the Statement of Need section as a whole addresses the question, and we believe it contains sufficient detail to provide a comprehensive response.

Since RStudio is simply an IDE on top of R, users working directly in R (e.g., via terminal) can follow the same installation steps, omitting only the RStudio installation.

Automated tests are included in the tests/testthat/ folder and can be run using devtools::test().

The acronyms in “Prepare your own input data” are either package names (BASD, STITCHES), model names (e.g., GFDL-ESM2M, HadGEM2-ES, IPSL-CM5A-LR, MIROC5), or project names (e.g., ISIMIP). Since we provide links to the relevant package and project pages, we believe users have enough context while keeping the instructions focused and concise.
