### 2025-09-29 4622c27
more minor wording updates; adding 2nd affiliations
