### 2025-09-29 b9b65c3
cont. fixes on affiliation
