### 2025-09-29 93baf06
proofread and more minor changes; addressing comments from Stephanie
