+++ paper/paper.md
@@ -4,13 +4,13 @@ authors:
 - affiliation: 1
   name: Di Sheng
   orcid: 0009-0006-7567-380X
-- affiliation: "1, 2"
+- affiliation: 1
   name: Xin Zhao
   orcid: 0000-0002-1801-4393
 - affiliation: 1
   name: Abigail Snyder
   orcid: 0000-0002-9034-9948
-- affiliation: "1, 2"
+- affiliation: 1
   name: Stephanie Morris
   orcid: 0000-0002-8073-0868
 - affiliation: 1
@@ -32,17 +32,15 @@ tags:
 affiliations:
 - index: 1
   name: Joint Global Change Research Institute, Pacific Northwest National Laboratory, College Park, MD, USA
-- index: 2
-  name: Center for Global Sustainability, School of Public Policy, University of Maryland, College Park, MD, USA
 ---
 
 # Summary
 
-Labor productivity, especially in outdoor sectors like primary agriculture, is highly sensitive to weather variations and heat stress. Understanding how labor productivity responds to key heat stress factors, such as temperature and humidity, is essential for evaluating future climate scenarios and their multisectoral dynamic (MSD) impacts. However, global economic and MSD modeling often operate at more aggregate temporal and spatial resolutions than Earth System Models. This resolution disparity highlights the need for tools to connect the fine-scale insights of Earth System Models with the more aggregate scales of economic models. The HELPS R package is designed to generate high-quality data on biophysical labor productivity losses due to heat stress, tailored for modeling applications. HELPS processes high-resolution atmospheric data to calculate heat stress metrics, such as Wet-Bulb Globe Temperature (WBGT) and translates them into physical work capacity (PWC) estimates across various scenarios. The package further provides functions to aggregate data temporally and spatially, incorporating crop calendars and harvested area information to provide relevant measures of heat-induced PWC losses for labor in agricultural sectors. We demonstrate the application of HELPS by generating regional heat-induced agricultural labor productivity losses for major crop sectors, which serve as input data for the Global Change Analysis Model (GCAM), enabling more comprehensive evaluations of climate impacts on agriculture systems. While HELPS is designed to support MSD modeling needs, users can customize output to benefit a broad set of research applications.
+Labor productivity, especially in outdoor sectors like primary agriculture, is highly sensitive to weather variations and heat stress. Understanding how labor productivity responds to key heat stress factors, such as temperature and humidity, is essential for evaluating future Earth system scenarios and their Multi-Sector Dynamic (MSD) impacts. However, global economic and MSD modeling often operate at more aggregate temporal and spatial resolutions than Earth System Models. This resolution disparity highlights the need for tools to connect the fine-scale insights of Earth System Models with the more aggregate scales of economic models. The HELPS R package is designed to generate high-quality data on biophysical labor productivity losses due to heat stress, tailored for modeling applications. HELPS processes high-resolution atmospheric data to calculate heat stress metrics, such as Wet-Bulb Globe Temperature (WBGT), and translates them into physical work capacity (PWC) estimates across various scenarios. The package further provides functions to aggregate data temporally and spatially, incorporating crop calendars and harvested area information to provide relevant measures of heat-induced PWC losses for agricultural labor. We demonstrate the application of HELPS by generating regional heat-induced agricultural labor productivity losses for major crop sectors, which serve as input data for the Global Change Analysis Model (GCAM), enabling more comprehensive evaluations of biophysical impacts on agriculture systems. While HELPS is designed to support MSD modeling needs, users can customize output to benefit a broad set of research applications.
 
 # Statement of need
 
[…]
