+++ paper/paper.bib
@@ -29,7 +29,8 @@
   number={1},
   pages={400},
   year={2025},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s43247-025-02318-w}
 }
 
 @article{sheng2025labour,
@@ -40,7 +41,8 @@
   number={2},
   pages={139--150},
   year={2025},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s43016-024-01088-6}
 }
 
 @article{casanueva2020escalating,
@@ -170,4 +172,16 @@
   year={2020},
   publisher={Annual Reviews},
   doi={10.1146/annurev-earth-053018-060100}
+}
+
+@article{calvin2019gcam,
+  title={GCAM v5. 1: representing the linkages between energy, water, land, climate, and economic systems},
+  author={Calvin, Katherine and Patel, Pralit and Clarke, Leon and Asrar, Ghassem and Bond-Lamberty, Ben and Cui, Ryna Yiyun and Di Vittorio, Alan and Dorheim, Kalyn and Edmonds, Jae and Hartin, Corinne and others},
+  journal={Geoscientific Model Development},
+  volume={12},
+  number={2},
+  pages={677--698},
+  year={2019},
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  doi={10.5194/gmd-12-677-2019}
 }
