+++ paper/paper.md
@@ -40,11 +40,11 @@ Labor productivity, especially in outdoor sectors like primary agriculture, is h
 
 # Statement of need
 
-Rising evidence of the adverse impact of heat stress on labor [@orlov2020economic], particularly for outdoor agricultural work [@de2021heat], underscores the critical need to quantify human heat stress exposure and heat-induced loss of PWC, particularly under hotter and more humid futures. Such quantification is essential for evaluating the MSD impact of biophysical shocks under different socioeconomic and Earth system scenarios and for guiding adaptation strategies to enhance resilience in vulnerable sectors, such as agriculture. We developed a code base to generate data for our paper studying heat-induced labor productivity loss and its implication for global agriculture[@shengomitting]. The HELPS R package, built upon the existing code base, provides a comprehensive tool for quantifying heat stress levels and the resulting PWC losses that is easy to use for global economic and MSD modeling.
+Rising evidence of the adverse impact of heat stress on labor [@orlov2020economic], particularly for outdoor agricultural work [@de2021heat], underscores the critical need to quantify human heat stress exposure and heat-induced loss of PWC, particularly under hotter and more humid futures. Such quantification is essential for evaluating the MSD impact of biophysical shocks under different socioeconomic and Earth system scenarios and for guiding adaptation strategies to enhance resilience in vulnerable sectors, such as agriculture. We developed a code base to generate data for our paper studying heat-induced labor productivity loss and its implication for global agriculture [@shengomitting]. The HELPS R package, built upon the existing code base, provides a comprehensive tool for quantifying heat stress levels and the resulting PWC losses that is easy to use for global economic and MSD modeling.
 
 To our knowledge, three existing R packages, HeatStress [@casanueva2020escalating], heatmetrics [@spangler2022wet], and meteor [@hijmans2023package], deliver related outputs. These packages provide functions to calculate heat stress levels using various heat stress metrics, i.e., WBGT, Universal Thermal Climate Index (UTCI), humidex, and heat index. The meteor package further translates heat stress to PWC using the labor-heat response functions from @smallcombe2022quantifying and @foster2021advanced. However, their estimates represent a single set of impacts that do not account for variations in labor heat exposure across crop types and management practices and need further processing to meet the spatial and temporal resolution demands of some research and modeling applications.
 
[…]
