### 2025-10-20 #8415 @dsheng1026
@anastassiavybornova Thank you, we have updated the paper and created a new release of the software (version 1.0.0). It is now archived on [zenodo](https://doi.org/10.5281/zenodo.17396826): https://doi.org/10.5281/zenodo.17396826
