### 2025-11-05 #8415 @dsheng1026
Hi, @anastassiavybornova I’d like to confirm whether JOSS requires a statement in the acknowledgments when an author’s current affiliation differs from the institution where the work was conducted. In my case, I am now affiliated with IIASA, while all work for this paper was completed at JGCRI (the affiliation listed in the manuscript).

Would it be appropriate to include a note such as:

“D.S. is currently affiliated with IIASA, which did not provide specific support for this paper.”

Thank you for clarifying.
