### 2025-11-05 #8415 @anastassiavybornova
@dsheng1026 looks like the 2 references above were defined twice in the bibliography, could you update and ping me once done? 🙏
