### 2025-11-05 e8be732
Update paper.bib
