+++ paper/paper.bib
@@ -152,16 +152,6 @@
   doi={10.5194/gmd-12-677-2019}
 }
 
-@article{sheng2025labour,
-  title={Labour market evolution is a key determinant of global agroeconomic and environmental futures},
-  author={Sheng, Di and Edmonds, James A and Patel, Pralit and Waldhoff, Stephanie T and O’Neill, Brian C and Wise, Marshall A and Zhao, Xin},
-  journal={Nature Food},
-  pages={1--12},
-  year={2025},
-  publisher={Nature Publishing Group UK London},
-  doi={10.1038/s43016-024-01088-6}
-}
-
 @article{buzan2020moist,
   title={Moist heat stress on a hotter Earth},
   author={Buzan, Jonathan R and Huber, Matthew},
@@ -172,16 +162,4 @@
   year={2020},
   publisher={Annual Reviews},
   doi={10.1146/annurev-earth-053018-060100}
-}
-
-@article{calvin2019gcam,
-  title={GCAM v5. 1: representing the linkages between energy, water, land, climate, and economic systems},
-  author={Calvin, Katherine and Patel, Pralit and Clarke, Leon and Asrar, Ghassem and Bond-Lamberty, Ben and Cui, Ryna Yiyun and Di Vittorio, Alan and Dorheim, Kalyn and Edmonds, Jae and Hartin, Corinne and others},
-  journal={Geoscientific Model Development},
-  volume={12},
-  number={2},
-  pages={677--698},
-  year={2019},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany},
-  doi={10.5194/gmd-12-677-2019}
 }
