### 2025-11-07 #8415 @kthyng
@dsheng1026 Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "earth" and "european".
