### 2025-11-07 269d616
Update paper.bib
