+++ paper/paper.bib
@@ -46,7 +46,7 @@
 }
 
 @article{casanueva2020escalating,
-  title={Escalating environmental summer heat exposure—a future threat for the European workforce},
+  title={Escalating environmental summer heat exposure—a future threat for the {E}uropean workforce},
   author={Casanueva, Ana and Kotlarski, Sven and Fischer, Andreas M and Flouris, Andreas D and Kjellstrom, Tord and Lemke, Bruno and Nybo, Lars and Schwierz, Cornelia and Liniger, Mark A},
   journal={Regional Environmental Change},
   volume={20},
@@ -98,13 +98,15 @@
   doi={10.1007/s00484-021-02105-0}
 }
 
-@article{jagermeyr2021ggcmi,
-  title={GGCMI Phase 3 crop calendar},
-  author={J{\"a}germeyr, Jonas and M{\"u}ller, Christoph and Minoli, Sara and Ray, Deepak and Siebert, Stefan},
-  journal={(No Title)},
-  year={2021},
-  publisher={Zenodo},
-  doi={10.5281/zenodo.5062513}
+@dataset{jagermeyr2021ggcmi,
+  title        = {GGCMI Phase 3 crop calendar},
+  author       = {J{\"a}germeyr, Jonas and M{\"u}ller, Christoph and Minoli, Sara and Ray, Deepak and Siebert, Stefan},
+  year         = {2021},
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.5062513},
+  url          = {https://doi.org/10.5281/zenodo.5062513},
+  version      = {1.0},
+  type         = {Dataset}
 }
 
 @misc{SPAM2024,
@@ -153,7 +155,7 @@
 }
 
 @article{buzan2020moist,
-  title={Moist heat stress on a hotter Earth},
+  title={Moist heat stress on a hotter {E}arth},
   author={Buzan, Jonathan R and Huber, Matthew},
   journal={Annual Review of Earth and Planetary Sciences},
   volume={48},
