### 2025-11-07 d03021b
Update paper.bib
