+++ paper/paper.bib
@@ -21,8 +21,8 @@
   doi={10.1088/1748-9326/abeb9f}
 }
 
-@misc{shengomitting,
-	title={Omitting labor responses underestimates the effects of future heat stress on agriculture},
+@article{shengomitting,
+  title={Omitting labor responses underestimates the effects of future heat stress on agriculture},
   author={Sheng, Di and Zhao, Xin and Edmonds, James A and Morris, Stephanie T and Patel, Pralit and O’Neill, Brian C and Tebaldi, Claudia and Wise, Marshall A},
   journal={Communications Earth \& Environment},
   volume={6},
