### 2025-11-07 #8415 @dsheng1026
@kthyng Thank you, and I just updated the .bib file.
