### 2025-11-07 #8415 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
